+++ src/routes/playlist/[id]/+page.svelte
@@ -1,5 +1,5 @@
 <script>
-  import { fetchApi, DeleteDialog, PlaylistStories, PlaylistMeta, PlaylistHeader  } from '$lib/index';
+  import { DeleteDialog, PlaylistStories, PlaylistMeta, PlaylistHeader  } from '$lib/index';
 
   export let data;
   let playlist = data?.playlist;
@@ -8,36 +8,14 @@
 
   $: if (playlist) isLoading = false;
 
-  let isLiked = playlist?.isLiked || false;
-  let existingLikeId = playlist?.likeId || null;
-  let profileId = 122;
-
-  async function toggleLike(event) {
-    event.preventDefault();
-    const endpoint = isLiked ? `/tm_likes/${existingLikeId}` : '/tm_likes';
-    const method = isLiked ? 'DELETE' : 'POST';
-
-    try {
-      const response = await fetchApi(endpoint, method, {
-        playlist: playlist.id,
-        profile: profileId
-      });
-
-      isLiked = !isLiked;
-      existingLikeId = response?.id || null;
-      playlist = { ...playlist, isLiked, likeId: existingLikeId };
-    } catch (err) {
-      console.error('Error toggling like:', err);
-      error = err.message || 'Er is iets fout gegaan';
-    }
-  }
 
 </script>
 
 <main>
   <article>
     <PlaylistHeader {playlist} />
-    <PlaylistMeta {playlist} {isLiked} {toggleLike} />
+    <!-- <PlaylistMeta {playlist} {isLiked} {toggleLike} /> -->
+     <PlaylistMeta {playlist} isLiked={playlist.isLiked} />
 
     {#if isLoading}
       <div class="loading">Loading playlist...</div>
